persists all console content. Re-order tabs. Change tab names. Moved console content management to the scheduler.

+++ android/src/com/google/wireless/speed/speedometer/BatteryCapPowerManager.java
@@ -79,7 +79,7 @@ public class BatteryCapPowerManager {
     private void broadcastMeasurementEnd(MeasurementResult result) {
       Intent intent = new Intent();
       intent.setAction(UpdateIntent.MEASUREMENT_PROGRESS_UPDATE_ACTION);
-      intent.putExtra(UpdateIntent.TASK_PRIORITY_PAYLOAD, realTask.getDescription().priority);
+      intent.putExtra(UpdateIntent.TASK_PRIORITY_PAYLOAD, (int) realTask.getDescription().priority);
       // A progress value MEASUREMENT_END_PROGRESS indicates the end of an measurement
       intent.putExtra(UpdateIntent.PROGRESS_PAYLOAD, Config.MEASUREMENT_END_PROGRESS);
       if (result != null) {
